Not sending sourceItemNumber anymore for orders and refunds.
I had mistakenly added sourceItemNumber to the parameters that get sent to Taxdoo for every item. It was sending the same ID that is sent for channelItemNumber, since I didn't have a complete understanding of the difference between the two. Since that's actually not a required field (if channelItemNumber is specified) and since there's not ERP integration, that field is completely unnecessary. Therefore I'm just dropping it.