Explicitly stating not to force the syncing in automatic sync.
The function isSyncable() of the class Order gives the option to force syncing, even if Transaction Sync has been disabled by the user.
This is useful when there's a backfill request, that should bypass the Transaction Sync option.
For the sake of clarity, I modified here two calls to the function isSyncable() to explicitly state NOT to force the syncing.