Removing the possibility to use the PUT method for updated orders/credit memos.
When an order or credit memo gets saved and is in a complete/closed state, the extension would verify if it has already been synced, and if it has been updated since it has been synced.
If it's not synced yet, it will sync it with the POST method to the resources /orders or /refunds.
In the case it has already been synced, but there have been updates, it would try to use the PUT method, to modify the existing transaction.
Problem is: Taxdoo won't allow the PUT method for the resources /orders and /refunds, but only for the /transactions resources, which is the most low-level one.
The required implementation is therefore quite different, and I decided to leave it out at the moment.
According to the Taxdoo API documentation, there's normally no need to update a past order, since refunds are handled in their own way.