Modifying the README to instruct the user to make "state" always mandatory
The Taxdoo API requires for every address a complete State (province, region, ...) field. Since Magento by default doesn't make the "State" field mandatory at checkout for all countries, I'm instructing the user of this extension to make the "State" field mandatory for every country, or at least every EU country. Otherwise if the customer decides not to fill that field, the order sync request will be refused by Taxdoo as a bad request.

This is of course suboptimal and could be automatically done using a plugin. This stackexchange response could be a good blueprint for a future implementation: https://magento.stackexchange.com/questions/169581/how-to-make-the-second-address-field-required-on-checkout-page-on-magento-2