Inserting a notification when a transaction fails to sync.
In case there's a failed attempt to sync an order or refund on Taxdoo, (a notification of the NotifierInterface kind) will appear in the admin dashboard.
The notification contains the order/refund number, the full text of the payload sent to Taxdoo, and the Taxdoo server response.
This can of course happen because of a bad API token (in this case a system message should already be present) or a wrong request.

Compare the server response with this list of status code: https://dev.taxdoo.com/#http-status-codes