Corrected the itemPrice value for the Taxdoo request to be multiplied by quantity
According to Taxdoo's documentation the itemPrice value "equals to the price per item multiplied with the quantity."
This wasn't the case in the code, since itemPrice was just the unit price of the object.
This has now been corrected both for orders and refunds.
The tests have been adjusted accordingly.