Complying with Magento 2 coding standard
Modifying the code to be compliant (or as compliant as possible) with Magento2 coding standards.
It consist mostly in indenting, adapting multiline strings/calls, spaces, and so on - and at times adding an Ignore when I couldn't think of an alternative.
Guilty as charged of including here also a correction to a small mistake I made in the logger, which was logging the wrong refund/order number.