Adding some coding standard ignores
Adding a couple of coding standard ignores to the code.
In one case it's about the obfuscateEmail function, that I implemented as a static one, after moving it to the helper.
In the other case it's about the Form class, which extends the Form class from the Magento framework. Its __construct method is an override which PHPCS judges useless. Since it's a class who stayed almost fully untouched from its TaxJar days, my understanding of it is still limited - and so is my understanding on the right way to listen to that warning :)