One more attempt at reducing cyclomatic complexity
The shouldSkipItem and buildLineItems contain many checks, and to reduce their cyclomatic complexity would require to split them more thoroughly.
At the moment I'm just suppressing the warning.
The same goes for the excessive number of dependencies of the Transaction and Backfill classes.

"Frenchmen, one more effort, if you want to be PHPMD compliant!"