Adding a testing class for Backfill
The Backfill class didn't have any unit or integration test, so we're filling that gap here.
Some refactoring was needed: that class used to have only one public method, laconically called start(), which did the following:
- get a start and end date as parameters (either from the CLI or the GUI)
- build a list of complete/closed orders between those two dates
- cycle through the orders and their creditmemos to build() them and push() them

Testing it as-it-was was inevitably a E2E situation, since the client gets called, and there's just that public method. Why not, one might ask - but we're not at E2E yet :-)
An alternative would have been to use mocks for orderFactory, and then for orders, so that the cycle through orders and creditmemos just doesn't do anything.

I decided therefore to make the createListOfOrders() method public, so that we can actually test it. At the end of the day, that's the main question about this class: is it backfilling the orders that it should backfill, and only those?
It wasn't straightforward to decide so, since that method is only called internally - but I thought that it could actually happen, that we'll want to use that method without actually doing the backfill, maybe to implement a dry run mode: "We're about to backfill the following orders, are you okay with that?". Therefore, it can make sense to make the method public, and therefore directly testable.

At the end of the day, a similar design was opted for by TaxJar in the Order and Refund classes, by splitting the build() and push() methods: the first one builds the payload, the second one pushes it remotely. They don't call each other (one could easily imagine push() calling a private/protected build() to create its payload), and when they're called in production code, they're called manually one after the other. The only case in which build() alone is called is in tests: that's why I suppose they were split for testing purposes.